Recalculate task index whenever a redraw occurs.
This makes sense, as a redraw only occurs if the list of tasks
changes, which implies that the index (i.e. their numeric order) also
changes -- especially when windows are destroyed/removed.

Previous to this, closing a window would cause a gap in numbering;
e.g. closing the 4th in 5 windows would leave 4 windows numbered
1,2,3,5 instead of the (originally) 5th window being renumbered.